+++ firmware/config/engines/mazda/mazda_miata_1_6.cpp
@@ -280,9 +280,6 @@ void setMiataNA6_MAP_MRE() {
 
 	engineConfiguration->fuelPumpPin = Gpio::Unassigned;
 
-
-	engineConfiguration->twoWireBatchInjection = true;
-
 	engineConfiguration->useIacTableForCoasting = true;
 	engineConfiguration->idlePidDeactivationTpsThreshold = 90;