+++ firmware/controllers/settings.cpp
@@ -594,9 +594,6 @@ static void enableOrDisable(const char *param, bool isEnabled) {
 		engineConfiguration->verboseKLine = isEnabled;
 	} else if (strEqualCaseInsensitive(param, "stepperidle")) {
 		engineConfiguration->useStepperIdle = isEnabled;
-	} else if (strEqualCaseInsensitive(param, "two_wire_batch_injection")) {
-		engineConfiguration->twoWireBatchInjection = isEnabled;
-		incrementGlobalConfigurationVersion();
 	} else if (strEqualCaseInsensitive(param, "boardUseTempPullUp")) {
 		engineConfiguration->boardUseTempPullUp = isEnabled;
 		incrementGlobalConfigurationVersion();