+++ firmware/CHANGELOG.md
@@ -35,6 +35,9 @@ Release template (copy/paste this for new release):
  - fuel_pump_control is accessible as getOutput("isFuelPumpOn") in Lua #5239 
  - start button logic does not crank started for first X seconds since ECU boot #5324
 
+### Breaking Changes
+ - Always operate in "two wire" mode for batch fuel, fixing batch firing order #5353
+
 ## June 2023 Release "Day 466"
 
 ### Breaking Changes